requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.1
fastapi==0.128.8
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
h11==0.16.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.12.0.88
packaging==26.0
pillow==11.3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
starlette==0.49.3
sympy==1.14.0
torch==2.2.2
torchvision==0.17.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
zipp==3.23.0
streamlit==1.50.0
torch
torchvision
pillow